Add files via upload
For creating the ML model, I used Support Vector model. The data set contained about 700 data points and the data points were standardized using StandardScaler library from sklearn.. After training the model with the data I got the accuracy score for test data as 77.27% and the accuracy score for train data displayed as 78.67%, which indicates that the model hasn't been over trained.  As a conclusion the model was able to correctly classify a diabetes patient and a non-diabetes patient.