Add files via upload
I used a SVM to build this Machine learning model. First the data was preprocessed before feeding into the model. The visualizations displayed that Males who are married and are Graduates with good credit score are more likely to receive a loan. For the visualization, I used seaborn library. After training the model the accuracy score for training data displayed as 77.08% and the accuracy score for test data displayed as 81.25%.